TINKERPOP-1004 Expand usage of TransactionException
TransactionException is no longer bound to usage by AbstractTransaction. Precribed a more general usage of it with the Transaction interface.
