+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/structure/util/AbstractTransaction.java
@@ -101,12 +101,8 @@ public abstract class AbstractTransaction implements Transaction {
     @Override
     public void commit() {
         readWrite();
-        try {
-            doCommit();
-            fireOnCommit();
-        } catch (TransactionException te) {
-            throw new RuntimeException(te);
-        }
+        doCommit();
+        fireOnCommit();
     }
 
     /**
@@ -115,12 +111,8 @@ public abstract class AbstractTransaction implements Transaction {
     @Override
     public void rollback() {
         readWrite();
-        try {
-            doRollback();
-            fireOnRollback();
-        } catch (TransactionException te) {
-            throw new RuntimeException(te);
-        }
+        doRollback();
+        fireOnRollback();
     }
 
     /**
@@ -152,23 +144,5 @@ public abstract class AbstractTransaction implements Transaction {
     @Override
     public void close() {
         doClose();
-    }
-
-    /**
-     * An "internal" exception thrown by providers when calls to {@link AbstractTransaction#doCommit} or
-     * {@link AbstractTransaction#doRollback} fail.
-     */
-    public static class TransactionException extends Exception {
-        public TransactionException(final String message) {
-            super(message);
-        }
-
-        public TransactionException(final Throwable cause) {
-            super(cause);
-        }
-
-        public TransactionException(final String message, final Throwable cause) {
-            super(message, cause);
-        }
     }
 }
